+++ app/routes/_index.tsx
@@ -1,5 +1,6 @@
 import type { MetaFunction } from "@remix-run/node";
-import WaitlistForm from "~/components/WaitList-Form";
+
+import WaitlistForm from "~/components/WaitListForm";
 
 export const meta: MetaFunction = () => {
   return [
@@ -11,8 +12,8 @@ export const meta: MetaFunction = () => {
 export default function Index() {
   return (
     <>
-    <h1>hello world</h1>
-    <WaitlistForm/>
+      <h1>hello world</h1>
+      <WaitlistForm />
     </>
   );
 }
